requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.3.0 # pyup: != 11.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==25.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.9.0  # https://github.com/evansd/whitenoise
redis==6.2.0  # https://github.com/redis/redis-py
hiredis==3.2.1  # https://github.com/redis/hiredis-py
celery==5.5.3  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.8.1  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower
uvicorn[standard]==0.35.0  # https://github.com/encode/uvicorn
uvicorn-worker==0.3.0  # https://github.com/Kludex/uvicorn-worker

# Django
# ------------------------------------------------------------------------------
django==5.1.11  # pyup: < 5.2 # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.9.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.4  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2025.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==6.0.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.16.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.7.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.28.0  # https://github.com/tfranzel/drf-spectacular
# DRF JWT
djangorestframework_simplejwt==5.5.0
